pyproject.toml
[tool.poetry]
name = "sarvam assignment"
version = "0.1.0"
description = ""
authors = ["jkumarr10 <jitthenkumarr29@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.3.2"
pandas = "^2.2.3"
dspy-ai = "^2.5.3"
langchain-community = "^0.3.1"
pypdf = "^5.0.1"
langchain-openai = "^0.2.2"
langchain-chroma = "^0.1.4"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
streamlit = "^1.39.0"
langgraph = "^0.2.34"
nest-asyncio = "^1.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
